feat(utils): convert getWidthByUpcount to TS, use theme import instead of function arg
